Convert position match from if statements to list of entries
